requirements.txt
annotated-types==0.6.0
anyio==3.7.1
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
dnspython==2.4.2
doqu==0.28.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.105.0
functions==0.7.0
greenlet==3.0.2
h11==0.14.0
httptools==0.6.1
idna==3.6
iniconfig==2.0.0
jwt==1.3.1
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.1
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pytest==7.4.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
schemas==0.7.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.9.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
